refactor: extract checkFields to service
