+++ src/lib/services/submit.services.ts
@@ -24,6 +24,7 @@ import { replaceHistory } from '$lib/utils/route.utils';
 import { isNullish } from '@dfinity/utils';
 import { getDoc, setDoc, type Doc } from '@junobuild/core-peer';
 import { nanoid } from 'nanoid';
+import { fileURLToPath } from 'url';
 
 export const initUserProposal = async ({
 	user,
@@ -390,3 +391,67 @@ const assertTimestamps = async () => {
 		);
 	}
 };
+
+export const checkFields = (
+	nodeProviderName: string,
+	url: string,
+	urlSelfDeclaration: string,
+	hashSelfDeclaration: string,
+	urlIdentityProof: string,
+	hashIdentityProof: string,
+	nodeProviderId: string
+  ): boolean => {
+	const validUrlDomain = 'https://wiki.internetcomputer.org/';
+	const sha256Regex = /^[a-fA-F0-9]{64}$/;
+	const fields = [
+	  nodeProviderName,
+	  url,
+	  urlSelfDeclaration,
+	  hashSelfDeclaration,
+	  urlIdentityProof,
+	  hashIdentityProof,
+	  nodeProviderId
+	];
+  
+	if (fields.some((field) => field === '' || field === undefined)) {
+	  toasts.error({
+		msg: { text: 'Please fill in all fields' }
+	  });
+	  return false;
+	}
+  
+	if (!urlSelfDeclaration.startsWith(validUrlDomain)) {
+	  toasts.error({
+		msg: {
+		  text: 'Invalid URL for self-declaration: make sure the url is from <https://wiki.internetcomputer.org/>'
+		}
+	  });
+	  return false;
+	}
+  
+	if (!urlIdentityProof.startsWith(validUrlDomain)) {
+	  toasts.error({
+		msg: {
+		  text: 'Invalid URL for identity proof: make sure the url is from <https://wiki.internetcomputer.org/>'
+		}
+	  });
+	  return false;
+	}
+  
+	if (!sha256Regex.test(hashSelfDeclaration)) {
+	  toasts.error({
+		msg: { text: 'Invalid hash for self-declaration: make sure this is a SHA256 hash' }
+	  });
+	  return false;
+	}
+  
+	if (!sha256Regex.test(hashIdentityProof)) {
+	  toasts.error({
+		msg: { text: 'Invalid hash for proof of identity: make sure this is a SHA256 hash' }
+	  });
+	  return false;
+	}
+  
+	return true;
+  };
+
